rename sprite animation component
