+++ doc/examples/animations/lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
 import 'package:flame/sprite_animation.dart';
-import 'package:flame/components/animation_component.dart';
+import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flutter/material.dart';
 
 void main() async {
