+++ doc/examples/render_flip/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flame/sprite_animation.dart';
-import 'package:flame/components/animation_component.dart';
+import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
 import 'package:flutter/material.dart';
