+++ doc/examples/spritesheet/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flame/components/animation_component.dart';
+import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flame/components/sprite_component.dart';
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
