+++ doc/examples/tiled/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flame/sprite_animation.dart';
-import 'package:flame/components/animation_component.dart';
+import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flame/components/tiled_component.dart';
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
