+++ lib/components/sprite_animation_component.dart
